+++ src/vue/pages/KycRecovery/KycRecoveryManagement.vue
@@ -14,7 +14,7 @@
           {{ 'kyc-recovery-state-message.sign-out-btn' | globalize }}
         </button>
       </div>
-      <h3>
+      <h3 class="kyc-recovery-management__functionality-msg">
         {{ 'kyc-recovery-state-message.limited-functionality-msg' | globalize }}
       </h3>
       <template v-if="isLoaded">
@@ -144,5 +144,9 @@ export default {
 
 .kyc-recovery-management__loader {
   margin-top: 2rem;
+}
+
+.kyc-recovery-management__functionality-msg {
+  color: $col-text;
 }
 </style>